Added the data tier and updated the pom xml with postgresql dependency, hope it works

+++ BankSystem/src/main/java/TransferedObjects/User.java
@@ -8,7 +8,7 @@ public class User implements Serializable {
     private UserTypes type;
     private double balance;
 
-    public User(String username,String password, UserTypes type, int balance)
+    public User(String username,String password, UserTypes type, double balance)
     {
         this.username = username;
         this.password = password;
